Bugfix: Allow backtracking after parsing a wrong character, fixes bug when dict in array contained only integers
Example:
[{"first":123,"second":234},{"first":123,"second":234}]

In this case the number parsing ate the closing curly bracket and the dict parsing failed because of a new
opening curly bracket inside a already started dict where a string was expected. Sounds complicated but
really isn't, we just have to re-parse that closing curly bracket in the dict stage.